fix more warnings

+++ src/lib/jp2/t1/T1Encoder.cpp
@@ -21,8 +21,8 @@
 
 namespace grk {
 
-T1Encoder::T1Encoder(grk_tcp *tcp, grk_tcd_tile *tile, uint16_t encodeMaxCblkW,
-		uint16_t encodeMaxCblkH, bool needsRateControl) :
+T1Encoder::T1Encoder(grk_tcp *tcp, grk_tcd_tile *tile, uint32_t encodeMaxCblkW,
+		uint32_t encodeMaxCblkH, bool needsRateControl) :
 		tile(tile),
 		needsRateControl(needsRateControl),
 		encodeBlocks(nullptr),
