+++ src/lib/jp2/t1/T1Factory.cpp
@@ -31,8 +31,8 @@ namespace grk {
 
 T1Interface* T1Factory::get_t1(bool isEncoder,
 								grk_tcp *tcp,
-								uint16_t maxCblkW,
-								uint16_t maxCblkH) {
+								uint32_t maxCblkW,
+								uint32_t maxCblkH) {
 	bool isHT = tcp->isHT;
 	if (isHT)
 		return new t1_ht::T1HT(isEncoder, tcp, maxCblkW, maxCblkH);
